(fleet) split the daemon and package manager part of the installer
