+++ pkg/installer/packages/service/apm_inject_test.go
