+++ pkg/installer/packages/service/cmd_executor_windows.go
