+++ pkg/installer/packages/service/datadog_agent.go
@@ -13,7 +13,6 @@ import (
 	"os/exec"
 	"strings"
 
-	"github.com/DataDog/datadog-agent/pkg/util/installinfo"
 	"github.com/DataDog/datadog-agent/pkg/util/log"
 	"gopkg.in/DataDog/dd-trace-go.v1/ddtrace/tracer"
 )
@@ -83,10 +82,6 @@ func SetupAgentUnits(ctx context.Context) (err error) {
 			return
 		}
 	}
-	// write installinfo before start, or the agent could write it
-	if err = installinfo.WriteInstallInfo("updater_package", "manual_update"); err != nil {
-		return
-	}
 	for _, unit := range stableUnits {
 		if err = startUnit(ctx, unit); err != nil {
 			return
@@ -133,7 +128,6 @@ func RemoveAgentUnits(ctx context.Context) {
 	if err := rmAgentSymlink(ctx); err != nil {
 		log.Warnf("Failed to remove agent symlink: %s", err)
 	}
-	installinfo.RmInstallInfo()
 }
 
 // StartAgentExperiment starts the agent experiment
