+++ pkg/installer/packages/service/datadog_agent_windows.go
