+++ pkg/installer/packages/service/datadog_installer_windows.go
