+++ pkg/installer/packages/service/docker_test.go
