+++ pkg/installer/packages/service/helper/go.mod
@@ -1,0 +1,3 @@
+module github.com/DataDog/datadog-agent/pkg/installer/packages/service/helper
+
+go 1.21.0
