+++ pkg/installer/repository/repository.go
@@ -17,7 +17,7 @@ import (
 
 	"github.com/DataDog/gopsutil/process"
 
-	"github.com/DataDog/datadog-agent/pkg/installer/service"
+	"github.com/DataDog/datadog-agent/pkg/installer/packages/service"
 	"github.com/DataDog/datadog-agent/pkg/util/log"
 )
