--- pkg/installer/service/helper/go.mod
@@ -1,3 +1,0 @@
-module github.com/DataDog/datadog-agent/pkg/installer/service/helper
-
-go 1.21.0
