+++ tasks/installer.py
@@ -52,7 +52,7 @@ def build(
 
     helper_bin = os.path.join(BIN_PATH, bin_name("helper"))
     helper_ldflags = f"-X main.installPath={install_path} -w -s"
-    helper_path = os.path.join("pkg", "installer", "service", "helper")
+    helper_path = os.path.join("pkg", "installer", "packages", "service", "helper")
     cmd = f"CGO_ENABLED=0 go build {build_type} -tags \"{go_build_tags}\" "
     cmd += f"-o {helper_bin} -gcflags=\"{gcflags}\" -ldflags=\"{helper_ldflags}\" {helper_path}/main.go"
